Inline btc_get_balance_test
