--- rs/tests/src/btc_integration/mod.rs
@@ -1,1 +1,0 @@
-pub mod btc;
